# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CoursesList API
  version: "1.0.0"
  summary: API systemu obsługi szkoleń — kursy, kursanci, firmy, zaświadczenia, dzienniki szkoleń.
  description: |
    Specyfikacja dla aplikacji integrujących się z API serwer-serwer za pomocą klucza API.

    **Przeczytaj najpierw `INTEGRATION.md`** z tego samego katalogu — opisuje reguły, których
    nie da się wyrazić w samym schemacie (semantyka PATCH jako pełnego nadpisania, formaty dat
    bez strefy czasowej, niespójności nazw pól, pułapki przy tworzeniu zaświadczeń).

    Konwencje:
    - Każda udana odpowiedź JSON ma kopertę `{"data": ...}`; błędy mają `{"error": {"code", "message"}}`.
    - Każda operacja wymaga zakresu podanego w `x-required-scope` (brak = wystarczy ważny klucz).
    - Operacje z `x-requires-admin: true` wymagają dodatkowo, by konto serwisowe klucza miało rolę
      administratora (`role = 1`) — sam zakres nie wystarczy.
    - Ciała żądań są odrzucane (400), jeśli zawierają pole spoza schematu.

    Wygenerowano z kodu źródłowego (`api/internal/*`) na gałęzi `api_and_webhook`.
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: http://localhost:8081
        description: >-
          Adres Go API. Lokalnie `http://localhost:8081`; w `docker-compose.prod.yml` API nasłuchuje na
          porcie 8082. Integracje powinny wołać Go API bezpośrednio, nie przez proxy Nuxt.
security:
  - apiKey: []
...
paths:
...
  /courses:
    get:
      tags: [courses]
      operationId: listCourses
      summary: Lista kursów (bez treści)
      description: >-
        `search` dopasowuje fragment w nazwie głównej, nazwie lub symbolu. Wyniki są rankingowane:
        dokładny symbol, symbol zaczynający się od frazy, nazwa zaczynająca się od frazy, nazwa
        główna zaczynająca się od frazy, symbol zawierający frazę, reszta; w obrębie grupy
        sortowanie po symbolu i nazwie.
      x-required-scope: courses:read
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Kursy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [courses]
      operationId: createCourse
      summary: Utworzenie kursu
      x-required-scope: courses:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseWrite"
      responses:
        "201":
          description: Utworzony kurs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetailsResponse"
        "400":
          description: >-
            `invalid request body` (brak wymaganego pola, ujemne `expiryTime`), `course program must be a JSON array`,
            `unsupported translation language`, `duplicate translation language`, `translation fields are required`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: "Kurs o tym symbolu już istnieje (`failed to create course: symbol exist`)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: >-
        Klucz API w nagłówku `Authorization: Bearer clk_...`. Klucze wystawia administrator w aplikacji
        webowej (Administracja → Klucze API). Klucz działa w imieniu przypisanego konta serwisowego i
        tylko w zakresie nadanych uprawnień (`x-required-scope`).
  parameters:
...
    Search:
      name: search
      in: query
      description: Fraza wyszukiwania; białe znaki na brzegach są obcinane, pusta = brak filtra.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maksymalna liczba wyników. Poza zakresem 1–100 → 400 `incorrect limit value`.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    BadRequest:
      description: >-
        Błędne żądanie — np. `invalid request body` (brak wymaganego pola, pole spoza schematu,
        zła data), błędny identyfikator w ścieżce, `incorrect limit value`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: >-
        Brak lub nieważny klucz: `missing credentials`, `invalid or expired api key`.
        Klucz mógł zostać unieważniony, wygasnąć albo jego konto usunięto.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: >-
        Klucz jest ważny, ale nie ma uprawnień: `api key is missing the <zakres> scope`
        albo `admin access required` (konto serwisowe nie jest administratorem).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Błąd serwera (`internal_error`). Treść komunikatu nie jest stabilna.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, invalid_credentials, forbidden, not_found, conflict, internal_error]
            message:
              type: string
              description: >-
                Komunikat po angielsku. Komunikaty wymienione w opisach operacji są stabilne i można na nich
                opierać logikę; pozostałe traktuj jako informacyjne.
...
    CourseListItem:
      type: object
      required: [id, mainName, name, symbol, expiryTime]
      properties:
        id:
          type: integer
          format: int64
        mainName:
          type: string
        name:
          type: string
        symbol:
          type: string
          description: Unikalny symbol kursu, używany w numerze zaświadczenia.
        expiryTime:
          $ref: "#/components/schemas/ExpiryTime"
    CourseListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CourseListItem"
    ExpiryTime:
      type: [integer, "null"]
      minimum: 0
      description: >-
        Okres ważności zaświadczenia w **latach**. `null` = kurs bez terminu ważności; `0` to poprawna wartość
        (ważność kończy się z końcem kursu). Data wygaśnięcia = `courseDateEnd` + `expiryTime` × 365 dni.
      examples: [5]
    CourseProgram:
      type: string
      contentMediaType: application/json
      description: >-
        **String zawierający JSON** (nie zagnieżdżony obiekt) — tablicę pozycji programu:
        `[{"Subject": "Temat", "TheoryTime": "4", "PracticeTime": "0"}]`. Klucze wielką literą,
        godziny jako stringi z liczbą. Program jest źródłem sesji dziennika i tabeli na wydruku.
      examples:
        - '[{"Subject":"Przepisy BHP","TheoryTime":"4","PracticeTime":"0"}]'
    CertFrontPage:
      type: string
      description: >-
        Szablon HTML pierwszej strony zaświadczenia. Placeholdery w podwójnych nawiasach klamrowych
        (spacje wewnątrz dowolne): `{{ imie }}`, `{{ drugie_imie }}`, `{{ nazwisko }}`, `{{ pesel }}`,
        `{{ data_urodzenia }}`, `{{ miejsce_urodzenia }}`, `{{ nazwa_kursu }}`, `{{ data_rozpoczecia }}`,
        `{{ data_zakonczenia }}`, `{{ data_wystawienia }}`, `{{ numer_zaswiadczenia }}`.
        Wartości są wstawiane z escapowaniem HTML, daty w formacie polskim.
    CourseCertificateTranslation:
      type: object
      required: [languageCode, courseName, courseProgram, certFrontPage]
      properties:
        languageCode:
          $ref: "#/components/schemas/TranslationLanguageCode"
        courseName:
          type: string
          minLength: 1
        courseProgram:
          $ref: "#/components/schemas/CourseProgram"
        certFrontPage:
          $ref: "#/components/schemas/CertFrontPage"
    TranslationLanguageCode:
      type: string
      enum: [en, de, uk, cs, sk, lt]
      description: "Języki tłumaczeń. `pl` to język bazowy kursu i nie może być tłumaczeniem."
    CourseDetails:
      type: object
      required: [id, mainName, name, symbol, expiryTime, courseProgram, certFrontPage, certificateTranslations]
      properties:
        id:
          type: integer
          format: int64
        mainName:
          type: string
        name:
          type: string
        symbol:
          type: string
        expiryTime:
          $ref: "#/components/schemas/ExpiryTime"
        courseProgram:
          $ref: "#/components/schemas/CourseProgram"
        certFrontPage:
          $ref: "#/components/schemas/CertFrontPage"
        certificateTranslations:
          type: array
          description: Posortowane po kodzie języka; pusta tablica, gdy brak tłumaczeń.
          items:
            $ref: "#/components/schemas/CourseCertificateTranslation"
    CourseDetailsResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CourseDetails"
...
    CourseWrite:
      type: object
      additionalProperties: false
      required: [mainName, name, symbol, expiryTime, courseProgram, certFrontPage]
      properties:
        mainName:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        symbol:
          type: string
          minLength: 1
        expiryTime:
          type: integer
          minimum: 0
          description: Wymagane (nie może być `null`). Nieujemna liczba całkowita lat.
          examples: [5]
        courseProgram:
          $ref: "#/components/schemas/CourseProgram"
        certFrontPage:
          $ref: "#/components/schemas/CertFrontPage"
        certificateTranslations:
          type: array
          description: >-
            Pełna lista tłumaczeń. Przy PATCH: pominięcie pola zostawia tłumaczenia bez zmian, podana lista
            zastępuje istniejące (tłumaczenia spoza listy są usuwane, `[]` usuwa wszystkie). Kody języków nie mogą
            się powtarzać; wszystkie pola każdego tłumaczenia są wymagane, a `courseProgram` musi być tablicą JSON.
          items:
            $ref: "#/components/schemas/CourseCertificateTranslation"

    # ---------------------------------------------------------- certificates
...